Cleaner architecture
